YJIT: Introduce a Flags for some small cleanup
We have all these ci flags scattered about the code base, always
asking questions like is it a splat call or a kw call. I find it nicer
to just have these methods to call.

Originally I wanted to make it so we could just pass around as CI
structure and not need to have flags as a separate argument. But right
now with the structure of optimize calls, we modify the flags and so
that wasn't doable here.

I would love to continue refactoring this code to the point were we
could lessen the number of arguments here and not need to pass the
flags, this is an incremental step in that direction.

I'll admit that this isn't a huge improvement or a groundbreaking
change. It is a small change that I think moves us in a direction
towards having a slightly higher level of abstraction with less copy
and paste. I personally have dealt with a bug that was about != 0 vs
== 0 that I think would have been easier to split with this code
change.